Move useCreateAssignablePaymentMethods to own file
